refactor(filesystem): split mega lib module
